+++ src/main/java/com/belogrudov/javabot/data/UsersTable.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsersTable extends JpaRepository<User, Long> {
+public interface UsersTable extends JpaRepository<User, Integer> {
     Boolean existsByChatId(Long chatId);
 
     User findByChatId(Long chatId);
